refactor all interface, dao and add alias to mybatis
