+++ src/main/java/com/solvd/page_rank/dao/IBaseDAO.java
@@ -1,4 +1,4 @@
-package com.solvd.page_rank.interfaces;
+package com.solvd.page_rank.dao;
 
 import java.util.List;
