+++ src/main/java/com/solvd/page_rank/dao/IPagesToRankDAO.java
@@ -1,5 +1,6 @@
-package com.solvd.page_rank.interfaces;
+package com.solvd.page_rank.dao;
 
+import com.solvd.page_rank.dao.IBaseDAO;
 import com.solvd.page_rank.models.PagesToRank;
 
 public interface IPagesToRankDAO extends IBaseDAO<PagesToRank> {
