+++ src/main/java/com/solvd/page_rank/dao/dao/mybatis/AbstractDAO.java
@@ -1,4 +1,4 @@
-package com.solvd.page_rank.dao;
+package com.solvd.page_rank.dao.dao.mybatis;
 
 import com.solvd.page_rank.myBatis.MyBatisSQLFactory;
 import org.apache.ibatis.session.SqlSession;
