+++ src/main/resources/myBatis/mappers/SettingForAlgorythmMapper.xml
@@ -2,29 +2,29 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.solvd.page_rank.interfaces.ISettingsForAlgorythmDAO">
-    <resultMap id="settingForAlgorythmResultMap" type="com.solvd.page_rank.models.SettingsForAlgorythm">
+<mapper namespace="com.solvd.page_rank.dao.ISettingsForAlgorythmDAO">
+    <resultMap id="settingForAlgorythmResultMap" type="SettingsForAlgorythm">
         <id property="id" column="id" javaType="int"/>
         <result property="dempingFactor" column="demping_factor"/>
         <result property="limitOfDeflect" column="limit_of_deflect"/>
-        <association property="user" column="user_id" javaType="com.solvd.page_rank.models.Users"
-                     select="com.solvd.page_rank.interfaces.IUsersDAO.getEntity"/>
+        <association property="user" column="user_id" javaType="Users"
+                     select="com.solvd.page_rank.dao.IUsersDAO.getEntity"/>
     </resultMap>
 
     <select id="getEntity" resultMap="settingForAlgorythmResultMap">
         SELECT * FROM settingforalgorythm WHERE id = #{id}
     </select>
 
-    <delete id="deleteEntity" parameterType="com.solvd.page_rank.models.SettingsForAlgorythm">
+    <delete id="deleteEntity" parameterType="SettingsForAlgorythm">
         DELETE FROM settingforalgorythm WHERE id = #{id}
     </delete>
 
-    <insert id="createEntity" useGeneratedKeys="true" keyProperty="id" parameterType="com.solvd.page_rank.models.SettingsForAlgorythm">
+    <insert id="createEntity" useGeneratedKeys="true" keyProperty="id" parameterType="SettingsForAlgorythm">
         INSERT INTO settingforalgorythm (demping_factor, limit_of_deflect, user_id)
         VALUES (#{dempingFactor}, #{limitOfDeflect}, #{user.id})
     </insert>
 
-    <update id="updateEntity" parameterType="com.solvd.page_rank.models.SettingsForAlgorythm">
+    <update id="updateEntity" parameterType="SettingsForAlgorythm">
         UPDATE settingforalgorythm SET demping_factor = #{dempingFactor}, limit_of_deflect = #{limitOfDeflect},
         user_id = #{user.id}
         WHERE id = #{id}
