+++ src/main/resources/myBatis/mappers/UsersMapper.xml
@@ -2,21 +2,21 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.solvd.page_rank.interfaces.IUsersDAO">
-    <resultMap id="usersResultMap" type="com.solvd.page_rank.models.Users">
+<mapper namespace="com.solvd.page_rank.dao.IUsersDAO">
+    <resultMap id="usersResultMap" type="Users">
         <id property="id" column="id" javaType="int"/>
         <result property="login" column="login"/>
         <result property="password" column="password"/>
         <collection property="pagesToRanks" javaType="ArrayList" column="id"
-                    ofType="com.solvd.page_rank.models.PagesToRank" select="getPagesToRank"/>
+                    ofType="PagesToRank" select="getPagesToRank"/>
     </resultMap>
-    <resultMap id="pagesToRankResultMap" type="com.solvd.page_rank.models.PagesToRank">
+    <resultMap id="pagesToRankResultMap" type="PagesToRank">
         <id property="id" column="id" javaType="int"/>
         <result property="rank" javaType="double" column="rank"/>
-        <association property="user" column="user_id" javaType="com.solvd.page_rank.models.Users"
-                     select="com.solvd.page_rank.interfaces.IUsersDAO.getEntity"/>
-        <association property="page" column="page_id" javaType="com.solvd.page_rank.models.Pages"
-                     select="com.solvd.page_rank.interfaces.IPagesDAO.getEntity"/>
+        <association property="user" column="user_id" javaType="Users"
+                     select="com.solvd.page_rank.dao.IUsersDAO.getEntity"/>
+        <association property="page" column="page_id" javaType="Pages"
+                     select="com.solvd.page_rank.dao.IPagesDAO.getEntity"/>
     </resultMap>
 
     <select id="getEntity" resultMap="usersResultMap">
@@ -31,15 +31,15 @@
         SELECT * FROM pagestorank WHERE user_id = #{user_id}
     </select>
 
-    <delete id="deleteEntity" parameterType="com.solvd.page_rank.models.Users">
+    <delete id="deleteEntity" parameterType="Users">
         DELETE FROM users WHERE id = #{id}
     </delete>
 
-    <insert id="createEntity" useGeneratedKeys="true" keyProperty="id" parameterType="com.solvd.page_rank.models.Users">
+    <insert id="createEntity" useGeneratedKeys="true" keyProperty="id" parameterType="Users">
         INSERT INTO users (login, password) VALUES (#{login}, #{password})
     </insert>
 
-    <update id="updateEntity" parameterType="com.solvd.page_rank.models.Users">
+    <update id="updateEntity" parameterType="Users">
         UPDATE users SET login = #{login}, password = #{password}
         WHERE id = #{id}
     </update>
