+++ src/main/resources/myBatis/my_batis_configuration.xml
@@ -5,6 +5,13 @@
 <configuration>
     <properties resource="db.properties">
     </properties>
+    <typeAliases>
+        <typeAlias type="com.solvd.page_rank.models.Pages" alias="Pages"/>
+        <typeAlias type="com.solvd.page_rank.models.PagesToRank" alias="PagesToRank"/>
+        <typeAlias type="com.solvd.page_rank.models.RelationsBetweenPages" alias="RelationsBetweenPages"/>
+        <typeAlias type="com.solvd.page_rank.models.SettingsForAlgorythm" alias="SettingsForAlgorythm"/>
+        <typeAlias type="com.solvd.page_rank.models.Users" alias="Users"/>
+    </typeAliases>
     <environments default="development">
         <environment id="development">
             <transactionManager type="JDBC" />
